created stylesheet and dock component
