+++ ducky/app/screens/Analytics.js
@@ -1,8 +1,11 @@
-import React, { Component } from 'react';
-import { View, Text, StyleSheet, Button } from 'react-native';
-import { LinearGradient } from 'expo-linear-gradient';
-import { AntDesign, MaterialIcons, Feather, MaterialCommunityIcons, FontAwesome } from '@expo/vector-icons';
+// Analytics.js
 
+import React, { Component } from 'react';
+import { View, Text, Button } from 'react-native';
+import { LinearGradient } from 'expo-linear-gradient';
+import { AntDesign, Feather, MaterialCommunityIcons, FontAwesome } from '@expo/vector-icons';
+
+import styles from '../Styles';
 
 export class Analytics extends Component {
   render() {
@@ -27,55 +30,3 @@ export class Analytics extends Component {
 }
 
 export default Analytics;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#AEC9EB',
-    paddingTop: 100,
-    // alignItems: 'center',
-    // justifyContent: 'center',
-  },
-  h5_white: {
-    color: '#fbfbfb', 
-    fontSize: 45,
-    fontWeight: 'bold'
-  },
-  h5_yellow: {
-    color: '#f9e067',
-    fontSize: 45,
-    fontWeight: 'bold'
-  },
-  cards: {
-    width: 300,
-    height: 450,
-    borderRadius: 20,
-    margin: 10
-  },
-  h4_blue: {
-    color: '#6B8DB2',
-    fontStyle: 'normal',
-    fontSize: 25,
-    fontWeight: '600',
-  },
-  dock_container: {
-    width: 414, 
-    height: 100, 
-    backgroundColor: '#AEC9EB', 
-    position: 'absolute', 
-    bottom: 0,
-    flex: 1,
-    alignContent: 'center',
-    justifyContent: 'center'
-  },
-  dock: {
-    flexDirection: 'row',
-    justifyContent: 'space-evenly',
-  },
-  h4_white: {
-    color: '#FBFBFB',
-    fontStyle: 'normal',
-    fontSize: 40,
-    fontWeight: 'bold',
-  }
-});
